# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RestAll API",
        "description": "REST API documentation for the RestAll restaurant management system.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get list of reservations for the day",
                "operationId": "f10e68ff0ce263ef754b2f640eda8d93",
                "responses": {
                    "200": {
                        "description": "List of reservations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "table_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "reservation_time": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-04-22T19:30:00Z"
                                            },
                                            "guests_count": {
                                                "type": "integer",
                                                "example": 4
                                            },
                                            "table": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "number": {
                                                        "type": "string",
                                                        "example": "T-1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Reserve a free table",
                "operationId": "4f59911f11563f79b4f01bd02222607d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "table_id",
                                    "reservation_time"
                                ],
                                "properties": {
                                    "table_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "reservation_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-04-22T19:30:00Z"
                                    },
                                    "guests_count": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "table_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "reservation_time": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-04-22T19:30:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "409": {
                        "description": "Conflict (Table is not available)"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/{id}": {
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Cancel a reservation",
                "operationId": "3f3a1d6e7ffee88b7ef58128a7f9295c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Reservation cancelled successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the token obtained during login",
                "scheme": "bearer"
            }
        }
    },
...
}